extract methods which are to be used by module_store into common parent class
